Update customFixes.fix
Removed the DEBAX, BIBER, TAKEK, and PUYAF fixes (lines 6-9) as they are very far outside of CZUL's borders. In fact the first three are about 1600nm south of Montreal, and the last one is 400nm west of Montreal, in the heart of Chicago ARTCC.

Added FAWNS to the list. FAWNS is a very common and frequently used fix when passing traffic to Boston Centre.